consolidate requirement to RAScript code in single class; split AchievementBuilder into smaller parts
