+++ Source/Parser/AchievementScriptInterpreter.cs
@@ -343,7 +343,7 @@ namespace RATools.Parser
 
             foreach (var leaderboard in _leaderboards.Keys)
             {
-                var leaderboardMinimumVersion = AchievementBuilder.GetMinimumVersion(leaderboard);
+                var leaderboardMinimumVersion = LeaderboardBuilder.GetMinimumVersion(leaderboard);
                 if (leaderboardMinimumVersion > minimumVersion)
                     minimumVersion = leaderboardMinimumVersion;
             }
